requirements.txt
slack_bolt
requests
mysql.connector